Proyecto final/backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.119.1",
    "psycopg[binary]>=3.2.11",
    "python-dotenv>=1.1.1",
    "uvicorn>=0.38.0",
]
